Create intro page skeleton
A button and an image, that's all there is folks! This commit sets up
that intro page according to the wireframes.  Eventually the image will
need to be created, though it won't be finalized until the rest of the
app is created (since it will need to reference visual components in the
instructions)

Issue #1 Crate Intro View