Refactor ViewControllers
Now that we have multiple view controllers I would like them to be
better organized.

This also includes some changes to the way the storyboard was set up,
and the correction of an event that apparently got re-inserted during
some merge resolution somewhere.

Needless to say, I'm looking forward to no longer working in the
storyboard, since WYSIWYG is chaos and makes it nearly impossible to
properly organize commits.