Pass data to intro view
The intro view is now given the model controller.  Currently we are
having it generate the script when the view is loaded.  I suspect this
will not be the place we ultimately want to keep it, since we want to
regenerate the script every time an interview is completed, but this is
fine for now.

Issue #13 Set up dynamic passage rendering
Issue #14 Set up dynamic question script