Create interview protocol
The interview protocol is applied to ViewControllers that are part of
the interview storyboard.  Really the point here is that there is an
interview model controller which needs to be passed allong throughout
the interview interfaces, so that the interfaces are appropriately
rendering (and tracking) various questions.

Eventually this model is probably going to be passing around the
recording object as well.

All this commit does is set up the protocol and apply it to the view
controllers.  The actual passing around of the populated instance will
come next.

Issue #13 Set up dynamic passage rendering
Issue #14 Set up dynamic question script