Pass around the interview model
We want the interview model to be injected into the view controllers
throughout the entire interview process.  This allows the interview
script to be defined at the beginning of the screen flow, and reset in
very specific and controlled situations.

Issue #13 Set up dynamic passage rendering
Issue #14 Set up dynamic question script