Load questions as blocks
As described in Issue #20 we want to load questions in blocks, rather
than as a giant set of individual items.  This allows the script to have
a bit more control over the order of questions, as desired.  In order to
make the tool more generally useful I decided to have it still load in
questions from the base question directory, and just treat that as
a "Block 0."

This commit also does some refactoring around how content is loaded,
creating a few helper methods to support things like listing items in
a path, extracting content from a file, etc.  At some point this may be
long in a utility object but for now I'm OK leaving it in here.

Issue #20 Add question groups