Fix block loading
The block loading script isn't perfectly safe; that should be fixed up
if we stick with this approach.  In the short term, however, we just
have to be sure to initialize the element at index 0 so that it is
empty, this way if there are no root level questions the blocks will
properly populate starting at block index 1.