Generate question script
Behold, questions are now selected from the question pool!

The logic is as follows: iterate through each block, selecting a random
question from each.  When a block is empty, skip it and continue to the
next block.  Repeat this until all questions have been selected.

It is highly unlikely that a user would go through the entire script in
a single session, but this generates the entire possible script up front
rather than generating it over time.  This seemed like an easier way to
keep track of state.

Note: this also fixes a bug where question pools were not being loaded
in alphabetical order, which meant that blocks were not appearing in the
correct order.

Issue #14 Set up dynamic question script