Enable recording
Recording what the user says is the primary purpose of this application.
Unfortunately there isn't any fancy pants cocoapod for recording like
there is for audio playback, and as a result we have to create something
to handle that ourselves.

This is just a partial implementation of the recording logic; in
particular it does not properly create multiple sets of recordings for
when the user hits replay, or goes back to earlier questions.
Furthermore, it does not time the total amount of material recorded.

It should be noted that originally I wanted to have the recording only
start after the question audio was finished; unfortunately it looks like
either the CocoaPod for SwiftySound is not quite up to the latest
version, or the callback-on-sound-completion functionality is broken.
Either way, after time boxing an attempt to debug this I decided to move
on for now.

Issue #17 Recording Module