Wait until audio stops to record
We do not want to record when the question is being read to the user.
This fixes that by using SwiftySound's event handler for sound
completion.

One note: it turns out that sound objects will fall out of scope before they fully
resolve, which is why can't just create the sound within that method and
stick with it.

Issue #30 Only record after the question audio is finished