Add participantId to models
Participant ID is a new field (a string) which will be entered by the
interviewer before passing things over to the participant.  This commit
doesn't actually make it possible to enter the data, but it sets up the
models for when they do.

Issue #29 Add ability to enter a participant ID