Track time recorded
The logic for this system is that each question will contribute a total
of MAX(answers(i)) seconds to the total recorded time.  For instance, if
the user answers the same question two times, one for 30 seconds one for
60 seconds, then it would only contribute the 60 second time to the
total.

In addition to this logic, we have inserted the "stop asking questions
after five minutes of quality answers" logic.

Issue #31 Implement 5 minute timer logic