Store versions of recordings
This commit redefines the way recorded responses are stored, making it
possible to store more than one response to the same question while also
encoding the id of the passage / question as part of the filename.

Under this logic there are two patterns for filenames.

For passage recordings:
{participant ID}_passage-v{response version}_{passage ID}_{timestamp}.m4a

For question recordings:
{participant ID}_question-{index}-v{response version}_{question
ID}_{timestamp}.m4a

Where:

- Participant ID is the value entered by the interviewer before starting
the interview.
- Response Version is the "take" (so if they repeat a question / go back
to a past question, it would start a new recording with an incremented
version).
- Passage / Question ID is the filename of the passage / question that
was rendered to the user.
- Timestamp is the unix timestamp at the moment that the recording
began.

Issue #37 Correct logic for repeating questions