Fix recording on physical devices
I was not properly setting up the shared session before starting
a recording.  This means that the recording did not work correctly when
kicked off on a physical iPad.  It isn't quite clear to me why it worked
in simulation but so it goes.

Another mysterious issue was that the very first question would not
properly get recorded (or time tracked).  This was solved by making sure
to re-set the category before each record() command.  I'm wary that this
is currently a fragile setup, but it does seem to work now.

Issue #41 Fix recording on physical iPad bug