Revamp passage recording logic
In the previous interview flow, a random passage was selected for
reading for each participant.  This reflects a change in that logic, so
that now all possible passages are presented to the user for reading.

In order to support this we needed to update the passage recording
controller to function more like the interview controller.
Specifically, the next and back buttons now need to iterate through the
list of selected passages and only continue to the next view in the
event that you are in the beginning (or end) of the passage list.

This also meant that we needed to add support for multiple versions of
a given passage recording, as users could now go "back" to re-record
a given passage.

This implementation is also more flexible for future in the event that
the logic is updated again; even if only one passage is "selected" the
passage recording controller will function in the same way.

Issue #45